LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/contain-intrinsic-size/auto-008-expected.txt


PASS requestAnimationFrame
FAIL Early ResizeObserver assert_equals: Using last remembered size - clientWidth expected 100 but got 40
FAIL Late ResizeObserver assert_equals: Using last remembered size - clientWidth expected 100 but got 40
